+++ CHANGELOG.md
@@ -16,6 +16,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
     5. Progress feedback and bar at the bottom of the main panel.
     6. Added formatted Table design for results.
 
+- Number of solutions found and whether if its Unique or not (single solution) - displayed on the statistics tab
+
 ### Changed
 
 ### Improved